requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
django-import-export==3.2.0
django-redis==5.3.0
et-xmlfile==1.1.0
frozenlist==1.3.3
heyoo==0.0.9
hiredis==2.2.3
idna==3.4
MarkupPy==1.14
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
prettytable==3.8.0
PyJWT==2.7.0
pymemcache==4.0.0
python-statemachine==2.1.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tabulate==0.9.0
texttable==1.6.7
tomli==2.0.1
transitions==0.9.0
twilio==8.3.0
typing==3.7.4.3
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
